fix file unzip

+++ build.py
@@ -71,10 +71,16 @@ def setup():
     extract_zip(os.path.join(DOWNLOAD_PATH, JETBRAINS_MONO_ZIP_NAME), DOWNLOAD_PATH)
     extract_zip(os.path.join(DOWNLOAD_PATH, JETBRAINS_MONO_NF_ZIP_NAME), DOWNLOAD_PATH)
 
-    print("[INFO] Deleting non-Regular ttf files.")
+    print("[INFO] Deleting non-Regular ttf files and other unnecessary files...")
     for root, dirs, files in os.walk(DOWNLOAD_PATH):
         for file in files:
             file_path = os.path.join(root, file)
+
+            # 파일명에 'D2Coding'이 포함된 ttf 파일은 삭제하지 않음
+            if file.endswith(".ttf") and "D2Coding" in file:
+                print(f"  - Keeping D2Coding font file: {file_path}")
+                continue
+
             if file.endswith(".ttf") and "Regular" not in file:
                 print(f"  - Deleting: {file_path}")
                 os.remove(file_path)
